third_party/blink/web_tests/platform/generic/fast/forms/input-type-change-expected.txt

Test the changing of an input type=TEXT to type=IMAGE, and vice versa. This test is to make sure that the height and width attributes are used for only IMAGE type.

Check metrics before the change:
PASS text.offsetWidth == 19 is false
PASS text.offsetHeight == 17 is false
PASS image.offsetWidth is 19
PASS image.offsetHeight is 17
Check metrics after the change:
PASS text.offsetWidth is 19
PASS text.offsetHeight is 17
PASS image.offsetWidth == 19 is false
PASS image.offsetHeight == 17 is false
PASS successfullyParsed is true

TEST COMPLETE
